B050: Added "Last Discovered" key and time functions (#155)
Added "Last Discovered" key to netscan assets
Changed all instances of time format to be time.RFC3339 (ISO time)
Added the isDate and toLocalDate exported JS functions that check if a date conforms to our format, and returns the date in the user's timezone, respectively.
Implemented the above mentioned functions in the asset info in asset view and asset list.
Renamed keys in network scan to make them more presentable.